pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "seaborn>=0.13.2",
    "streamlit>=1.46.1",
]
